@@ -11,6 +11,10 @@
     {
       "name": "Diglin (Derivative Work)",
       "homepage": "https://www.diglin.com"
+    },
+    {
+      "name": "Gyroscops (Derivative Work)",
+      "homepage": "https://gyroscops.com"
     }
   ],
   "autoload": {
@@ -20,11 +24,12 @@
   },
   "autoload-dev": {
     "psr-4": {
+      "spec\\Diglin\\Sylius\\ApiClient\\": "spec/",
       "Diglin\\Sylius\\ApiClient\\tests\\": "tests/"
     }
   },
   "require": {
-    "php": ">=7.1",
+    "php": ">=8.0",
     "psr/http-client": "^1.0",
     "psr/http-message": "^1.0",
     "php-http/httplug": "^1.1 || ^2.0",
@@ -33,12 +38,13 @@
     "php-http/message-factory": "^v1.0",
     "php-http/multipart-stream-builder": "^1.0",
     "php-http/client-implementation": "^1.0",
-    "symfony/expression-language": "^3.0|^4.0|^5.0"
+    "symfony/expression-language": "^3.0|^4.0|^5.0",
+    "webmozart/assert": "^1.10"
   },
   "require-dev": {
     "friendsofphp/php-cs-fixer": "^2.14",
-    "phpunit/phpunit": "^5.7",
-    "phpspec/phpspec": "^5.0",
+    "phpunit/phpunit": "^9.0",
+    "phpspec/phpspec": "^7.1",
     "symfony/yaml": "^4.2",
     "donatj/mock-webserver": "^2.0",
     "php-http/guzzle6-adapter": "^2.0"
