@@ -17,11 +17,11 @@ interface DownloadableResourceInterface
     /**
      * Downloads a resource by its code.
      *
-     * @param string $code Code of the resource
+     * @param string|int $code Code of the resource
      *
      * @throws HttpException
      *
      * @return StreamInterface
      */
-    public function download($code);
+    public function download($code): StreamInterface;
 }
