@@ -16,7 +16,7 @@ interface UpsertableResourceInterface
     /**
      * Creates a resource if it does not exist yet, otherwise updates partially the resource.
      *
-     * @param string $code code of the resource to create or update
+     * @param string|int $code code of the resource to create or update
      * @param array  $data data of the resource to create or update
      *
      * @throws HttpException if the request failed
@@ -24,5 +24,5 @@ interface UpsertableResourceInterface
      * @return int Status code 201 indicating that the resource has been well created.
      *             Status code 204 indicating that the resource has been well updated.
      */
-    public function upsert($code, array $data = []);
+    public function upsert($code, array $data = []): int;
 }
