@@ -16,11 +16,11 @@ interface UriGeneratorInterface
     /**
      * Generate an uri from a path, by adding host and port.
      *
-     * @param string $path            Path of the endpoint
-     * @param array  $uriParameters   List of the parameters to generate the endpoint
-     * @param array  $queryParameters List of the query parameters added to the endpoint
+     * @param string|\Stringable $path Path of the endpoint
+     * @param array  $uriParameters    List of the parameters to generate the endpoint
+     * @param array  $queryParameters  List of the query parameters added to the endpoint
      *
      * @return UriInterface
      */
-    public function generate($path, array $uriParameters = [], array $queryParameters = []);
+    public function generate(string|\Stringable $path, array $uriParameters = [], array $queryParameters = []);
 }
