@@ -1,262 +1,12 @@
 <?php
-/**
- * Diglin GmbH - Switzerland.
- *
- * @author      Sylvain Rayé <support at diglin.com>
- *
- * @category    SyliusApiClient
- *
- * @copyright   2020 - Diglin (https://www.diglin.com)
- */
 
-namespace Diglin\Sylius\ApiClient;
+declare(strict_types=1);
 
-use Diglin\Sylius\ApiClient\Api;
-use Diglin\Sylius\ApiClient\Client\AuthenticatedHttpClient;
-use Diglin\Sylius\ApiClient\Client\HttpClient;
-use Diglin\Sylius\ApiClient\Client\ResourceClient;
-use Diglin\Sylius\ApiClient\FileSystem\FileSystemInterface;
-use Diglin\Sylius\ApiClient\FileSystem\LocalFileSystem;
-use Diglin\Sylius\ApiClient\Pagination\PageFactory;
-use Diglin\Sylius\ApiClient\Pagination\ResourceCursorFactory;
-use Diglin\Sylius\ApiClient\Routing\UriGenerator;
-use Diglin\Sylius\ApiClient\Security\Authentication;
-use Diglin\Sylius\ApiClient\Stream\MultipartStreamBuilderFactory;
-use Diglin\Sylius\ApiClient\Stream\UpsertResourceListResponseFactory;
-use Http\Client\HttpClient as Client;
-use Http\Discovery\HttpClientDiscovery;
-use Http\Discovery\Psr17FactoryDiscovery;
-use Http\Message\RequestFactory;
-use Http\Message\StreamFactory;
-use Psr\Http\Message\RequestFactoryInterface;
-use Psr\Http\Message\StreamFactoryInterface;
+use Diglin\Sylius\ApiClient;
+
+trigger_deprecation('diglin/sylius-api-php-client', '2.0', 'The "%s" class is deprecated, use "%s" instead.', 'Diglin\\Sylius\\ApiClient\\SyliusClientBuilder', ApiClient\SyliusLegacyClientFactory::class);
 
 /**
- * Builder of the class SyliusClient.
- * This builder is in charge to instantiate and inject the dependencies.
- *
- * @author    Alexandre Hocquard <alexandre.hocquard@akeneo.com>
- * @author    Sylvain Rayé <support@diglin.com>
- * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ * @deprecated since diglin/sylius-api-php-client 2.0, use Diglin\Sylius\ApiClient\ApiClient\SyliusLegacyClientBuilder instead.
  */
-class SyliusClientBuilder implements SyliusClientBuilderInterface
-{
-    /** @var string */
-    protected $baseUri;
-
-    /** @var Client */
-    protected $httpClient;
-
-    /** @var RequestFactory */
-    protected $requestFactory;
-
-    /** @var StreamFactory */
-    protected $streamFactory;
-
-    /** @var FileSystemInterface */
-    protected $fileSystem;
-
-    /** @var Api\ApiAwareInterface[] */
-    protected $apiRegistry = [];
-
-    /** @var array */
-    protected $defaultHeaders = [];
-
-    public function __construct(?Api\ApiAwareInterface ...$apis)
-    {
-        foreach ($apis as $api) {
-            $this->addApi($api);
-        }
-    }
-
-    /**
-     * @return array
-     */
-    protected function setUp(Authentication $authentication)
-    {
-        $uriGenerator = new UriGenerator($this->baseUri);
-        $httpClient = new HttpClient($this->getHttpClient(), $this->getRequestFactory(), $this->getStreamFactory(), $this->defaultHeaders);
-
-        $authenticationApi = new Api\AuthenticationApi($httpClient, $uriGenerator);
-        $authenticatedHttpClient = new AuthenticatedHttpClient($httpClient, $authenticationApi, $authentication);
-        $multipartStreamBuilderFactory = new MultipartStreamBuilderFactory($this->getStreamFactory());
-        $upsertListResponseFactory = new UpsertResourceListResponseFactory();
-
-        $resourceClient = new ResourceClient(
-            $authenticatedHttpClient,
-            $uriGenerator,
-            $multipartStreamBuilderFactory,
-            $upsertListResponseFactory
-        );
-
-        $pageFactory = new PageFactory($authenticatedHttpClient, $uriGenerator);
-        $cursorFactory = new ResourceCursorFactory();
-        $fileSystem = null !== $this->fileSystem ? $this->fileSystem : new LocalFileSystem();
-
-        return [$resourceClient, $pageFactory, $cursorFactory, $fileSystem];
-    }
-
-    public function addApi(Api\ApiAwareInterface $api): self
-    {
-        $this->apiRegistry[(new \ReflectionClass($api))->getShortName()] = $api;
-
-        return $this;
-    }
-
-    public function setBaseUri(string $baseUri): SyliusClientBuilderInterface
-    {
-        $this->baseUri = $baseUri;
-
-        return $this;
-    }
-
-    public function setDefaultHeaders(array $headers): SyliusClientBuilderInterface
-    {
-        $this->defaultHeaders = $headers;
-
-        return $this;
-    }
-
-    /**
-     * Allows to directly set a client instead of using HttpClientDiscovery::find().
-     */
-    public function setHttpClient(Client $httpClient): SyliusClientBuilderInterface
-    {
-        $this->httpClient = $httpClient;
-
-        return $this;
-    }
-
-    /**
-     * Allows to directly set a request factory instead of using MessageFactoryDiscovery::find().
-     */
-    public function setRequestFactory(RequestFactory $requestFactory): SyliusClientBuilderInterface
-    {
-        $this->requestFactory = $requestFactory;
-
-        return $this;
-    }
-
-    /**
-     * Allows to directly set a stream factory instead of using StreamFactoryDiscovery::find().
-     */
-    public function setStreamFactory(StreamFactory $streamFactory): SyliusClientBuilderInterface
-    {
-        $this->streamFactory = $streamFactory;
-
-        return $this;
-    }
-
-    /**
-     * Build the Sylius client authenticated by user name and password.
-     */
-    public function buildAuthenticatedByPassword(
-        string $clientId,
-        string $secret,
-        string $username,
-        string $password
-    ): SyliusClientInterface {
-        $authentication = Authentication::fromPassword($clientId, $secret, $username, $password);
-
-        return $this->buildAuthenticatedClient($authentication);
-    }
-
-    /**
-     * Build the Sylius client authenticated by token.
-     */
-    public function buildAuthenticatedByToken(
-        string $clientId,
-        string $secret,
-        string $token,
-        string $refreshToken
-    ): SyliusClientInterface {
-        $authentication = Authentication::fromToken($clientId, $secret, $token, $refreshToken);
-
-        return $this->buildAuthenticatedClient($authentication);
-    }
-
-    /**
-     * Build the Sylius client authenticated by HTTP header.
-     */
-    public function buildAuthenticatedByHeader(array $xAuthToken): SyliusClientInterface
-    {
-        $authentication = Authentication::fromXAuthToken($xAuthToken);
-
-        return $this->buildAuthenticatedClient($authentication);
-    }
-
-    private function buildAuthenticatedClient(Authentication $authentication): SyliusClientInterface
-    {
-        [$resourceClient, $pageFactory, $cursorFactory, $fileSystem] = $this->setUp($authentication);
-
-        $client = new SyliusClientDecorator(
-            new SyliusClient(
-                $authentication,
-                new Api\CartsApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\ChannelsApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\CountriesApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\CurrenciesApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\CustomersApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\ExchangeRatesApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\LocalesApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\OrdersApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\PaymentMethodsApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\PaymentsApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\ProductsApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\ProductAttributesApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\ProductAssociationTypesApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\ProductOptionsApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\ProductReviewsApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\ProductVariantsApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\PromotionsApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\PromotionCouponsApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\ShipmentsApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\ShippingCategoriesApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\TaxCategoriesApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\TaxRatesApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\TaxonsApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\UsersApi($resourceClient, $pageFactory, $cursorFactory),
-                new Api\ZonesApi($resourceClient, $pageFactory, $cursorFactory)
-            )
-        );
-
-        foreach ($this->apiRegistry as $key => $api) {
-            $api->setResourceClient($resourceClient);
-            $api->setPageFactory($pageFactory);
-            $api->setCursorFactory($cursorFactory);
-            $api->setFileSystem($fileSystem);
-
-            $client->addApi($key, $api);
-        }
-
-        return $client;
-    }
-
-    private function getHttpClient(): Client
-    {
-        if (null === $this->httpClient) {
-            $this->httpClient = HttpClientDiscovery::find();
-        }
-
-        return $this->httpClient;
-    }
-
-    private function getRequestFactory(): RequestFactoryInterface
-    {
-        if (null === $this->requestFactory) {
-            $this->requestFactory = Psr17FactoryDiscovery::findRequestFactory();
-        }
-
-        return $this->requestFactory;
-    }
-
-    private function getStreamFactory(): StreamFactoryInterface
-    {
-        if (null === $this->streamFactory) {
-            $this->streamFactory = Psr17FactoryDiscovery::findStreamFactory();
-        }
-
-        return $this->streamFactory;
-    }
-}
+class_alias(ApiClient\SyliusLegacyClientFactory::class, 'Diglin\\Sylius\\ApiClient\\SyliusClientBuilder');
